+++ tests/integration/display-deep-model-test.js
@@ -36,7 +36,7 @@ module('Display deep model', {
       let commentTable = db['ember-flexberry-dummy-comment'];
       let commentVoteTable = db['ember-flexberry-dummy-comment-vote'];
       db.transaction('rw', suggestionTable, suggestionTypeTable, applicationUserTable, voteTable, commentTable, commentVoteTable,  function() {
-        suggestionTable.add({
+        suggestionTable.put({
           id: 'fea5b275-cb9b-4584-ba04-26122bc8cbd3',
           address: 'Street, 20',
           text: 'Loooong text',
@@ -51,7 +51,7 @@ module('Display deep model', {
           comments: ['7e5d3b63-eb5e-446e-84da-26865f87c1c5']
         });
 
-        suggestionTypeTable.bulkAdd([
+        suggestionTypeTable.bulkPut([
           {
             id: 'de627522-47c3-428f-99be-fdac2e8f5618',
             name: '123',
@@ -68,7 +68,7 @@ module('Display deep model', {
           }
         ]);
 
-        applicationUserTable.add({
+        applicationUserTable.put({
           id: '555a6d25-ac76-417c-bcc5-25bc260fc3ae',
           name: 'Васиииилий',
           eMail: 'pupkin1@mail.ru',
@@ -85,14 +85,14 @@ module('Display deep model', {
           karma: 11.4
         });
 
-        voteTable.add({
+        voteTable.put({
           id: '8be0d89b-8cab-4b0b-b029-356c59809163',
           suggestion: 'fea5b275-cb9b-4584-ba04-26122bc8cbd3',
           voteType: 'Like',
           applicationUser: '555a6d25-ac76-417c-bcc5-25bc260fc3ae'
         });
 
-        commentTable.add({
+        commentTable.put({
           id: '7e5d3b63-eb5e-446e-84da-26865f87c1c5',
           suggestion: 'fea5b275-cb9b-4584-ba04-26122bc8cbd3',
           text: 'Not ok',
@@ -102,7 +102,7 @@ module('Display deep model', {
           userVotes: ['721e65db-9e04-47a3-8f29-3b5c39fff8dd']
         });
 
-        commentVoteTable.add({
+        commentVoteTable.put({
           id: '721e65db-9e04-47a3-8f29-3b5c39fff8dd',
           comment: '7e5d3b63-eb5e-446e-84da-26865f87c1c5',
           voteType: 'Dislike',